backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.2
djangorestframework==3.16.0
python-dotenv==1.1.1
sqlparse==0.5.3
